Add files via upload
There is a pond. Inside the pond, there are 50 pebbles, arranged in a cycle. A frog is sitting on one of the pebbles. Whenever he jumps, he lands exactly 21 pebbles away in the clockwise direction. What is the minimum number of jumps in the clockwise direction the frog needs to reach its meal?